# Conflicts:
#	typos.toml

+++ src/layout/work.zig
@@ -9,7 +9,7 @@ const Ident = @import("base").Ident;
 /// Key to identify a type variable in a specific module.
 /// Used to distinguish type vars with the same index across different modules.
 pub const ModuleVarKey = packed struct {
-    module_idx: u16,
+    module_idx: u32,
     var_: types.Var,
 };
 
@@ -68,7 +68,7 @@ pub const Work = struct {
     /// A container being processed. The var_ is optional because synthetic tuples
     /// (created for multi-arg tag union variants) don't have a meaningful var to cache.
     /// module_idx tracks which module the var belongs to for correct in_progress_vars removal.
-    pub const PendingContainerItem = struct { var_: ?types.Var, module_idx: u16, container: PendingContainer };
+    pub const PendingContainerItem = struct { var_: ?types.Var, module_idx: u32, container: PendingContainer };
 
     /// Tuple field for layout work - similar to RecordField but with index instead of name.
     /// We need to explicitly record the index because zero-sized tuple fields might have
